Quick Refactor: Moving some function to common place.
